pyproject.toml
[tool.poetry]
name = "lnshop"
version = "0.2.0"
description = ""
authors = ["Kornpow <test@email.com>"]

[tool.poetry.dependencies]
python = ">3.6.1 <=3.8.6"
django = "^3.1.1"
django-rest-framework = "^0.1.0"
django-extensions = "^3.0.8"
django-cors-headers = "^3.5.0"
django-restql = "^0.10.2"
PyQRCode = "^1.2.1"
pypng = "^0.0.20"
schedule = "^0.6.0"
uvicorn = "^0.11.8"
gunicorn = "^20.0.4"
pillow = "^7.2.0"
websocket = "^0.2.1"
whitenoise = "5.2.0"
lnd-pyshell = {git = "https://github.com/sako0938/lnd_pyshell.git"}
django-filter = "^2.4.0"

[tool.poetry.dev-dependencies]
pytest = "^3.4"